[+] release configuration [+] bucket grid [+] network basic

+++ neurolife/neurolife/view_model.h
@@ -14,7 +14,7 @@ public:
 	virtual  ~BaseViewModel() {
 	}
 
-	double elsize() const;
+	int elsize() const;
 
 protected:
